Use PageSaver

+++ common/src/main/scala/org/overviewproject/models/Page.scala
@@ -24,7 +24,7 @@ object Page {
   case class ReferenceAttributes(
     id: Long,
     fileId: Long,
-    pageNumber: Long,
+    pageNumber: Int,
     text: String
   )
 }
